pre lib prep

+++ src/booster/IBooster.sol
@@ -2,12 +2,12 @@
 pragma solidity ^0.8.20;
 
 interface IBooster {
-    function percentageBoost(
-        address claimant,
-        uint baseFlax
-    ) external view returns (uint percentageBoost, uint sFlaxBalance);
+    function percentageBoost(address claimant, uint256 baseFlax)
+        external
+        view
+        returns (uint256 percentageBoost, uint256 sFlaxBalance);
 
-    function BasisPoints() external view returns (uint basisPoints);
+    function BasisPoints() external view returns (uint256 basisPoints);
 
-    function burnSFlax(address claimant, uint amount) external ;
+    function burnSFlax(address claimant, uint256 amount) external;
 }
